+++ tck/src/main/java/jakarta/json/bind/tck/defaultmapping/polymorphictypes/TypeInfoExceptionsTest.java
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package jakarta.json.bind.tck.defaultmapping.typeinfo;
+package jakarta.json.bind.tck.defaultmapping.polymorphictypes;
 
 import java.lang.invoke.MethodHandles;
