+++ tck/src/main/java/jakarta/json/bind/tck/defaultmapping/polymorphictypes/model/StringContainerSubClass.java
@@ -18,16 +18,16 @@
  * $Id$
  */
 
-package jakarta.json.bind.tck.defaultmapping.typeinfo.model;
+package jakarta.json.bind.tck.defaultmapping.polymorphictypes.model;
 
-public class StringContainer {
-  private String instance = "Test String";
+public class StringContainerSubClass extends StringContainer {
+  private String newInstance = "SubClass Test String";
 
-  public String getInstance() {
-    return instance;
+  public String getNewInstance() {
+    return newInstance;
   }
 
-  public void setInstance(String instance) {
-    this.instance = instance;
+  public void setNewInstance(String newInstance) {
+    this.newInstance = newInstance;
   }
 }
